Take the isolated mutation runner to a full mutation score
Its supervisor tests were sitting in a file the gate could not match to it,
so more than half of what they check was invisible: moving them into the
mirror folder alone took the score from 68.6% to 95.5%.

The rest are new checks — that the child is told it is running in a snapshot,
that its pid is printed, that the run record is written before the wait for
the lock, that a finished run is never signalled however alive its pid looks,
and that a stopping child gets its moment to end before being forced. Drops a
second write of the same record that the write above it had already done.

The isolation-state tests move to their own mirror folder as well, which is
where they belonged; that file can now be mutation-checked too.
